Backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
boto3==1.34.60
pdfplumber==0.11.0
pytesseract==0.3.10
pdf2image==1.17.0
PyMySQL==1.1.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0